git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/branches/ProbABEL-pacox/v.0.3.0/ProbABEL@1224 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ src/gendata.h
@@ -32,7 +32,7 @@ public:
             unsigned int npeople, unsigned int nmeasured,
             unsigned short int * allmeasured, std::string * idnames);
 
-    void get_var(int var, double * data);
+    void get_var(int var, float * data);
 
     ~gendata();
 
@@ -40,7 +40,7 @@ public:
     // ANOTHER PRIVATE OBJECT IS A POINTER TO DATABELBASECPP
     // UPDATE SNP, ALL REGRESSION METHODS: ACCOUNT FOR MISSING
 private:
-    mematrix<double> G;
+    mematrix<float> G;
     AbstractMatrix * DAG;
     unsigned short int * DAGmask;
 };
